+++ CHANGELOG.md
@@ -17,7 +17,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add redis and vt loading [32](https://github.com/greenbone/eulabeia/pull/32)
 - Possibility to preprocess messages [31](https://github.com/greenbone/eulabeia/pull/31)
 - Possibility to send one start.scan event containing all the data [31](https://github.com/greenbone/eulabeia/pull/31)
+- Add result model [36](https://github.com/greenbone/eulabeia/pull/36)
 - Add model for vts [37](https://github.com/greenbone/eulabeia/pull/37)
+
 ### Changed
 - Split cmds and info messages into own module [8](https://github.com/greenbone/eulabeia/pull/8)
 - Normalized topic structure to `group/aggregate/event/destination`; setting topic based on return message rather than configuration [8](https://github.com/greenbone/eulabeia/pull/8)